requirements.txt
# Reconhecimento de voz
speechrecognition
pyaudio

# API Gmail e autenticação Google
google-auth
google-auth-oauthlib~=1.2.2
google-api-python-client~=2.172.0

# Utilitários padrão
requests~=2.32.4
python-dotenv~=1.1.0


# Agente de IA Autogen
autogen-agentchat
autogen-core
autogen-ext

# Azure AI Client
azure-core~=1.34.0
azure-ai-ml
openai

# Processamento de texto
regex

# Necessário para execução do cliente e escuta de voz
keyboard
pynput~=1.8.1

# Áudio (caso queira manter suporte alternativo ao whisper ou sounddevice)
scipy
numpy~=2.2.6


# Tipagem
typing_extensions
openai-whisper~=20240930
sounddevice~=0.5.2
soundfile~=0.13.1
protobuf~=5.29.5
h11~=0.16.0
anyio~=4.9.0
pydantic~=2.11.5
httpx~=0.28.1
typing_extensions~=4.14.0
pydantic_core~=2.33.2
pip~=25.1.1
attrs~=25.3.0
distro~=1.9.0
cryptography~=45.0.4
filelock~=3.18.0
rsa~=4.9.1
pyasn1~=0.6.1
sniffio~=1.3.1
httpcore~=1.0.9
idna~=3.10
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.4.26
urllib3~=2.4.0
packaging~=25.0
cffi~=1.17.1
proto-plus~=1.26.1
six~=1.17.0
python-dateutil~=2.9.0.post0
pyparsing~=3.2.3
httplib2~=0.22.0
oauthlib~=3.2.2
Deprecated~=1.2.18
annotated-types~=0.7.0
pillow~=11.2.1
charset-normalizer~=3.4.2
Jinja2~=3.1.6
PyYAML~=6.0.2
setuptools~=80.9.0
uritemplate~=4.2.0
google-auth-httplib2~=0.2.0
requests-oauthlib~=2.0.0
pyautogen~=0.9.0
mcp~=1.9.4
httpx-sse~=0.4.0
typing_extensions~=4.14.0
starlette~=0.47.0
sse-starlette~=2.3.6
pydantic_core~=2.33.2
pydantic-settings~=2.9.1
click~=8.2.1
aiohttp~=3.12.12
pyasn1_modules~=0.4.2
uvicorn~=0.34.3
python-multipart~=0.0.20